Added build method for explore-request-builder
Explore API is now being used as part of other APIs too (like events) which do not inherently support many of the functionalities that explore does..
One requirement is to build the request using pre-defined scope and manipulating the arguments instead of using the user-entered arguments as is.
Hnece, added an overloaded build method which takes all the arguments as parameters.